refactor(traits): split traits.rs into focused modules
Break up the 2565-line traits.rs god object into smaller,
focused modules:

- timeout.rs: Timeout<T> utility for timed events
- symbols.rs: Common UI symbols and characters
- event.rs: EventResult, FocusStyle
- view.rs: View, Interactive, Draggable, StyledView traits
- widget_state.rs: WidgetState, WidgetProps, color resolution
- element.rs: Element enum
- render_context.rs: RenderContext with all drawing methods
- mod.rs: Re-exports and macro definitions

Closes #198 (partial)
